Do not delete .gitkeep files.

+++ build.mjs
@@ -42,6 +42,8 @@ async function cleanRealmAutoFiles(realm) {
 	)
 
 	for (const entry of entries) {
+		if (entry === ".gitkeep") continue
+
 		await fs.rm(
 			path.join("src", `realm-${realm}`, "auto", entry), {
 				recursive: true,
